pyproject.toml
[project]
name = "robot-vacuume-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-miio>=0.5.12",
    "uvloop>=0.21.0",
]
